% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline labels of earthquakes}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, n_max = NULL,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}. If specified and \code{inherit.aes = TRUE}
(the default), it is combined with the default mapping at the top level of
the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If \code{NULL}, the default, the data is inherited from the plot data as
specified in the call to \code{\link[ggplot2]{ggplot}}.
A \code{data.frame}, or other object, will override the plot data. All
objects will be fortified to produce a data frame. See
\code{\link[ggplot2]{fortify}} for which variables will be created.
A \code{function} will be called with a single argument, the plot data.
The return value must be a \code{data.frame}, and will be used as the layer
data.}

\item{...}{other arguments passed on to \code{layer}. These are often
aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters to the
paired geom/stat.}

\item{n_max}{An integer. If used, it only plots the labels for the
\code{n_max} largest earthquakes in the selected group in the timeline}
}
\description{
This geom plots timeline labels of earthquakes. It assumes that
\code{geom_timeline} was used to create the timelines
}
\details{
The function plots timeline labels of earthquakes based on cleaned
NOAA data. It should be used with combination with \code{geom_timeline}. The
required aesthetics for this geom is \code{label} that should contain
string for labeling each data point.
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
   filter(COUNTRY \%in\% c("GREECE", "ITALY"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE,
              y = COUNTRY,
              color = as.numeric(TOTAL_DEATHS),
              size = as.numeric(EQ_PRIMARY)
   )) +
   geom_timeline() +
   geom_timeline_label(aes(label = LOCATION_NAME), n_max = 5) +
   theme(
      plot.background = element_blank(),
      panel.background = element_blank(),
      legend.key = element_blank(),
      axis.title.y = element_blank(),
      axis.line.x = element_line(size = 1),
      axis.ticks.y = element_blank(),
      legend.position = "bottom"
   ) +
   labs(size = "Richter scale value", color = "# deaths")
}
}
